add py39 to ci, skip rst tests on windows
