--- pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-pytester_example_dir = examples
