+++ setup.cfg
@@ -38,3 +38,10 @@ ignore=E128
 
 [flake8]
 ignore=E128
+
+[tool:pytest]
+pytester_example_dir = examples
+
+filterwarnings =
+    ignore::FutureWarning:
+    ignore::DeprecationWarning:requests_cache
