+++ test/test_cache.py
@@ -8,18 +8,6 @@ import pytest
 
 import requests_cache
 
-from . import examples
-
-
-@pytest.fixture
-def base_args():
-    return ["-v", "--check-links", "--check-links-cache"]
-
-
-@pytest.fixture
-def memory_args(base_args):
-    return base_args + ["--check-links-cache-backend", "memory"]
-
 
 def assert_sqlite(testdir, name=None, tmpdir=None, exists=True):
     name = name or ".pytest-check-links-cache.sqlite"
@@ -59,7 +47,10 @@ def test_cache_expiry(testdir, base_args, cache_name, tmpdir):
     t3 = time.time()
     result.assert_outcomes(**expected)
 
-    assert t1 - t0 > t3 - t2, "cache did not make second run faster"
+    d0 = t1 - t0
+    d1 = t3 - t2
+
+    assert d0 > d1, "cache did not make second run faster"
 
     time.sleep(2)
 
@@ -68,7 +59,10 @@ def test_cache_expiry(testdir, base_args, cache_name, tmpdir):
     t5 = time.time()
     result.assert_outcomes(**expected)
 
-    assert t5 - t4 > t3 - t2, "cache did not expire"
+    d2 = t5 - t4
+    d3 = t3 - t2
+
+    assert d2 > d3, "cache did not expire"
 
 
 def test_cache_memory(testdir, memory_args):
